Upload files to "v2-gui/BOTWM_Autoupdater_GUI"

+++ v2-gui/BOTWM_Autoupdater_GUI/Form1.cs
@@ -15,6 +15,64 @@ namespace BotWMultiplayerUpdaterGUI
         private static readonly string releaseZipFile = "latest_release.zip";
         private static readonly string versionFileName = "Version.txt";
         private static readonly string updaterFileName = "BOTWM_Autoupdater_GUI.exe";
+        private static readonly HashSet<string> allowedFiles = new HashSet<string>
+        {
+            //Current Versions
+			"BOTWM_Autoupdater.exe",
+            "Breath of the Wild Multiplayer.deps.json",
+            "Breath of the Wild Multiplayer.dll",
+            "Breath of the Wild Multiplayer.exe",
+            "Breath of the Wild Multiplayer.runtimeconfig.json",
+            "Newtonsoft.Json.dll",
+            "Version.txt",
+            //Linux
+            "BOTWM_Autoupdater",
+			//Old Version
+			".gitignore",
+            "BOTW Multiplayer.exe",
+            "BOTW Multiplayer.exe.config",
+            "BOTW.DedicatedServer.deps.json",
+            "BOTW.DedicatedServer.dll",
+            "BOTW.DedicatedServer.exe",
+            "BOTW.DedicatedServer.runtimeconfig.json",
+            "BOTWM.Server.dll",
+            "BOTWMUpdater.py",
+            "Gamemodes.json",
+            "Microsoft.Bcl.AsyncInterfaces.dll",
+            "Microsoft.Bcl.AsyncInterfaces.xml",
+            "Newtonsoft.Json.dll",
+            "System.Buffers.dll",
+            "System.Buffers.xml",
+            "System.Memory.dll",
+            "System.Memory.xml",
+            "System.Numerics.Vectors.dll",
+            "System.Numerics.Vectors.xml",
+            "System.Runtime.CompilerServices.Unsafe.dll",
+            "System.Runtime.CompilerServices.Unsafe.xml",
+            "System.Text.Encodings.Web.dll",
+            "System.Text.Encodings.Web.xml",
+            "System.Text.Json.dll",
+            "System.Text.Json.xml",
+            "System.Threading.Tasks.Extensions.dll",
+            "System.Threading.Tasks.Extensions.xml",
+            "System.ValueTuple.dll",
+            "System.ValueTuple.xml",
+            "TeleportScript.py",
+            "version.txt"
+        };
+
+        private static readonly HashSet<string> allowedFolders = new HashSet<string>
+        {
+            //Current Versions
+			"Backgrounds",
+            "BNPs",
+            "DedicatedServer",
+			//Current and Old Versions
+			"Resources",
+			//Old Version
+			"BCML Files",
+            "libs"
+        };
 
         public Form1()
         {
@@ -47,6 +105,12 @@ namespace BotWMultiplayerUpdaterGUI
                 return;
             }
 
+            if (ContainsOtherFilesOrFolders())
+            {
+                labelStatus.Text = "Other files or folders detected. The updater will not run.";
+                return;
+            }
+
             string selectedVersion = listBoxVersions.SelectedItem.ToString();
             labelStatus.Text = $"Downloading version: {selectedVersion}...";
             DeleteFilesExceptUpdater();
@@ -76,20 +140,36 @@ namespace BotWMultiplayerUpdaterGUI
             }
         }
 
+        private bool ContainsOtherFilesOrFolders()
+        {
+            string folderPath = Directory.GetCurrentDirectory();
+            DirectoryInfo directory = new DirectoryInfo(folderPath);
+
+            var otherFiles = directory.GetFiles().Where(file => !allowedFiles.Contains(file.Name)).ToList();
+            var otherFolders = directory.GetDirectories().Where(dir => !allowedFolders.Contains(dir.Name)).ToList();
+
+            return otherFiles.Count > 0 || otherFolders.Count > 0;
+        }
+
         private void DeleteFilesExceptUpdater()
         {
             string folderPath = Directory.GetCurrentDirectory();
             DirectoryInfo directory = new DirectoryInfo(folderPath);
+
             foreach (FileInfo file in directory.GetFiles())
             {
-                if (file.Name != updaterFileName)
+                if (!allowedFiles.Contains(file.Name))
                 {
                     file.Delete();
                 }
             }
+
             foreach (DirectoryInfo dir in directory.GetDirectories())
             {
-                dir.Delete(true);
+                if (!allowedFolders.Contains(dir.Name))
+                {
+                    dir.Delete(true);
+                }
             }
         }
